requirements.txt
# ═══════════════════════════════════════════════════════════════════
# AI Trading System — Python Dependencies
# ═══════════════════════════════════════════════════════════════════
# pip install -r requirements.txt

# --- 核心依賴 ---
aiohttp>=3.9
numpy>=1.26
pandas>=2.1
pydantic>=2.0
apscheduler>=3.10
python-dotenv>=1.0
anthropic>=0.40

# --- 監控儀表板 ---
streamlit>=1.30
plotly>=5.18

# --- 開發 & 測試 ---
pytest>=8.0
pytest-asyncio>=0.23
aioresponses>=0.7
ruff>=0.3

# --- 選用：回測引擎 ---
# vectorbt>=0.26
yfinance>=0.2.36
